requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.0.1
cryptography==42.0.5
cssselect==1.2.0
cssutils==2.6.0
Django==4.1.7
django-phonenumber-field==7.0.2
django-storages==1.13.2
djangorestframework==3.14.0
google-api-core==2.11.0
google-auth==2.16.1
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
idna==3.4
lxml==4.9.2
phonenumbers==8.13.6
Pillow==9.4.0
premailer==3.10.0
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyMySQL==1.0.2
pytz==2022.7.1
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
yagmail==0.15.293
